moved and renamed the videos directory
